file rename & minor tweaks
